fix bugbot suggestions
