+++ src/main/kotlin/com/fcfb/discord/refbot/handlers/discord/DiscordMessageHandler.kt
@@ -901,23 +901,7 @@ class DiscordMessageHandler(
         offensiveCoaches: List<User?>,
         defensiveCoaches: List<User?>,
     ): Pair<Pair<String, EmbedData?>, List<User?>> {
-        val title =
-            when {
-                game.gameType == GameType.BOWL && game.bowlGameName?.isNotBlank() == true -> {
-                    "${game.homeTeam} vs ${game.awayTeam} | ${game.bowlGameName}"
-                }
-                game.gameType == GameType.CONFERENCE_CHAMPIONSHIP -> {
-                    val conferenceName = gameUtils.getConferenceName(game.homeTeam)
-                    if (conferenceName != null) {
-                        "${game.homeTeam} vs ${game.awayTeam} | $conferenceName Championship"
-                    } else {
-                        "${game.homeTeam} vs ${game.awayTeam} | Conference Championship"
-                    }
-                }
-                else -> {
-                    "${game.homeTeam} vs ${game.awayTeam}"
-                }
-            }
+        val title = gameUtils.getGameEmbedTitle(game)
         val embedData =
             EmbedData(
                 title = Optional(title),
@@ -957,23 +941,7 @@ class DiscordMessageHandler(
                 append("**" + game.awayTeam).append(":** ").append(game.awayScore).append("\n")
                 append("----------------\n")
             }
-        val title =
-            when {
-                game.gameType == GameType.BOWL && game.bowlGameName?.isNotBlank() == true -> {
-                    "${game.homeTeam} vs ${game.awayTeam} | ${game.bowlGameName}"
-                }
-                game.gameType == GameType.CONFERENCE_CHAMPIONSHIP -> {
-                    val conferenceName = gameUtils.getConferenceName(game.homeTeam)
-                    if (conferenceName != null) {
-                        "${game.homeTeam} vs ${game.awayTeam} | $conferenceName Championship"
-                    } else {
-                        "${game.homeTeam} vs ${game.awayTeam} | Conference Championship"
-                    }
-                }
-                else -> {
-                    "${game.homeTeam} vs ${game.awayTeam}"
-                }
-            }
+        val title = gameUtils.getGameEmbedTitle(game)
         val embedData =
             EmbedData(
                 title = Optional(title),
